ENH: add example folder
