+++ .gitignore
@@ -14,7 +14,6 @@
 
 # BIG TEST FILE #
 #################
-SRR494099.qscore
 SRR494099_big.qscore
 *SRR494099.txt
 *.subqscores
