+++ example/cleanup.py
@@ -1,0 +1,18 @@
+# cleanup.py
+
+import os
+
+from glob import glob
+from subprocess import call
+
+
+def main():
+    file_formats = ['bin', 'o', 'txt', 'subqscores']
+    for file_format in file_formats:
+        fnames = glob('./*.'+file_format)
+        for fname in fnames:
+            os.remove(fname)
+
+
+if __name__ == "__main__":
+    main()
